Added full ESRGAN-NCNN support, improved ShippedEsrgan installer

+++ Code/OS/ESRGAN.cs
@@ -18,15 +18,25 @@ namespace Cupscale.OS
 		public enum PreviewMode { None, Cutout, FullImage }
 		
 
-		public static async Task UpscaleBasic(string inpath, string outpath, ModelData mdl, string tilesize, bool alpha, PreviewMode mode, bool showTileProgress = true)
+		public static async Task UpscaleBasic(string inpath, string outpath, ModelData mdl, string tilesize, bool alpha, PreviewMode mode, bool allowNcnn, bool showTileProgress = true)
 		{
             try
             {
-				string modelArg = GetModelArg(mdl);
-				Logger.Log("Model Arg: " + modelArg);
-				Program.mainForm.SetProgress(5f, "Starting ESRGAN...");
-				await Run(inpath, outpath, modelArg, tilesize, alpha, showTileProgress);
-				File.Delete(Paths.progressLogfile);
+                if (allowNcnn && Config.GetBool("useNcnn"))
+                {
+					Program.mainForm.SetProgress(1f, "Loading ESRGAN-NCNN...");
+					Program.lastModelName = mdl.model1Name;
+					await RunNcnn(inpath, outpath, mdl.model1Path);
+				}
+                else
+                {
+					Program.mainForm.SetProgress(5f, "Starting ESRGAN...");
+					File.Delete(Paths.progressLogfile);
+					string modelArg = GetModelArg(mdl);
+					Logger.Log("Model Arg: " + modelArg);
+					await Run(inpath, outpath, modelArg, tilesize, alpha, showTileProgress);
+				}
+				
 				if (mode == PreviewMode.Cutout)
 				{
 					Program.mainForm.SetProgress(100f, "Merging into preview...");
@@ -50,6 +60,7 @@ namespace Cupscale.OS
 			catch(Exception e)
             {
 				MessageBox.Show("An error occured during upscaling: \n\n" + e.Message, "Error");
+				Logger.Log("Upscaling Error: " + e.Message + "\n" + e.StackTrace);
             }
 			
 		}
@@ -81,8 +92,8 @@ namespace Cupscale.OS
 
 		public static async Task Run(string inpath, string outpath, string modelArg, string tilesize, bool alpha, bool showTileProgress)
 		{
-			inpath = "\"" + inpath + "\"";
-			outpath = "\"" + outpath + "\"";
+			inpath = inpath.WrapPath();
+			outpath = outpath.WrapPath();
 			string alphaStr = " --noalpha";
 			if (alpha)
 				alphaStr = "";
@@ -157,39 +168,57 @@ namespace Cupscale.OS
 			await Task.Delay(1);
 		}
 
-		public static async Task RunNcnn (string inpath, string outpath, string modelArg, string tilesize)
+		public static string currentNcnnModel = "";
+
+		public static async Task RunNcnn (string inpath, string outpath, string modelPath)
 		{
-			/*
-			inpath = "\"" + inpath + "\"";
-			outpath = "\"" + outpath + "\"";
-			string alphaStr = " --noalpha";
-			if (alpha)
-				alphaStr = "";
-			string deviceStr = " --device cuda";
-			if (Config.GetBool("useCpu"))
-				deviceStr = " --device cpu";
-			string cmd2 = "/C cd /D \"" + Config.Get("esrganPath") + "\" & ";
-			cmd2 = cmd2 + "python esrlmain.py " + inpath + " " + outpath + deviceStr + " --tilesize " + tilesize + alphaStr + modelArg;
+			inpath = inpath.WrapPath();
+			outpath = outpath.WrapPath();
+			
+			Program.mainForm.SetProgress(3f, "Converting NCNN model...");
+			await NcnnUtils.ConvertNcnnModel(modelPath);
+			Logger.Log("NCNN Model is ready: " + currentNcnnModel);
+			Program.mainForm.SetProgress(5f, "Loading ESRGAN-NCNN...");
+			int scale = NcnnUtils.GetNcnnModelScale(currentNcnnModel);
+			string cmd2 = "/C cd /D \"" + Config.Get("esrganPath") + "\" & "
+				+ "esrgan-ncnn-vulkan.exe -i " + inpath + " -o " + outpath + " -m " + currentNcnnModel.WrapPath() + " -s " + scale;
 			Logger.Log("CMD: " + cmd2);
-			Process esrganProcess = new Process();
-			esrganProcess.StartInfo.UseShellExecute = false;
-			esrganProcess.StartInfo.RedirectStandardOutput = true;
-			esrganProcess.StartInfo.RedirectStandardError = true;
-			esrganProcess.StartInfo.CreateNoWindow = true;
-			esrganProcess.StartInfo.FileName = "cmd.exe";
-			esrganProcess.StartInfo.Arguments = cmd2;
-			esrganProcess.OutputDataReceived += OutputHandler;
-			esrganProcess.ErrorDataReceived += OutputHandler;
-			currentProcess = esrganProcess;
-			esrganProcess.Start();
-			esrganProcess.BeginOutputReadLine();
-			esrganProcess.BeginErrorReadLine();
-			while (!esrganProcess.HasExited)
+			Process ncnnProcess = new Process();
+			ncnnProcess.StartInfo.UseShellExecute = false;
+			ncnnProcess.StartInfo.RedirectStandardOutput = true;
+			ncnnProcess.StartInfo.RedirectStandardError = true;
+			ncnnProcess.StartInfo.CreateNoWindow = true;
+			ncnnProcess.StartInfo.FileName = "cmd.exe";
+			ncnnProcess.StartInfo.Arguments = cmd2;
+			ncnnProcess.OutputDataReceived += NcnnOutputHandler;
+			ncnnProcess.ErrorDataReceived += NcnnOutputHandler;
+			currentProcess = ncnnProcess;
+			ncnnProcess.Start();
+			ncnnProcess.BeginOutputReadLine();
+			ncnnProcess.BeginErrorReadLine();
+			while (!ncnnProcess.HasExited)
 			{
 				await Task.Delay(100);
 			}
-			File.Delete(Paths.progressLogfile);
-			*/
+		}
+
+		private static void NcnnOutputHandler(object sendingProcess, DataReceivedEventArgs output)
+		{
+			if (output == null || output.Data == null)
+				return;
+
+			string data = output.Data;
+			if (data.Contains("failed"))
+			{
+				if (currentProcess != null && !currentProcess.HasExited)
+					currentProcess.Kill();
+
+				MessageBox.Show("Error occurred: \n\n" + data + "\n\nThe ESRGAN-NCNN process was killed to avoid lock-ups.", "Error");
+				Logger.Log("Error Output: " + data.Replace("\n", " ").Replace("\r", " "));
+			}
+
+			if (data.Contains("vkAllocateMemory"))
+				MessageBox.Show("ESRGAN-NCNN ran out of memory. Try reducing the tile size and avoid running programs in the background (especially games) that take up your VRAM.", "Error");
 		}
 	}
 }
